libcontent: make content load/store API private
Problem: the public API functions for loading and storing blobs
in the content store would appear to encourage their use, but there
are caveats that make this probably not a great idea at this time.

The main caveat is the fact that content not linked to the KVS could
disappear across an instance checkpoint/restart.

Relocate the content functions from the public facing Flux API
to libcontent, and then include that library in libflux-internal.la.
Drop the flux_ prefixes from the functions and update callers.

Drop flux_content_load(3) man page.

Fixes #4213