broker: restrict content access to instance owner
Problem: guests can directly access the content store, but this
is both unnecessary and a possible avenue for a DoS attack, since
there are no restrictions on what can be stored.

In addition, guest access enables hash-guessing as a way to attempt
to obtain unauthorized access to job information, however unlikely
or impractical that means of attack may be.

Change message handlers for content.load and content.store to deny
access to guests.  Update sharness test.

Note this restriction in flux-content(1).