sched-simple: propagate any instance expiration into R
Problem: The current Flux instance time limit (as set in resource.R
`expiration` key) is currently ignored by sched-simple, but it would
be better to propagate the instance expiration to child jobs that
are submitted without duration, so that jobs do not appear as if
they have no time limit when one will be enforced a higher level in
the hierarchy.

If the current resource set has an expiration, ensure this value is
propagated to the R generated for jobs in sched-simple for all job
submitted with the default duration of 0.