job-manager: add builtin job duration validator
Problem: Jobs that are submitted with a duration that exceeds the
current instance lifetime should be rejected, but currently there is
no validation of the duration set in jobspec beyond the basic jobspec
validator.

Add a new builtin jobtap plugin which fetches the current instance R
and rejects jobs where the duration will exceed the current instance
lifetime if R contains an expiration timestamp. The plugin places a
kvs watch on resource.R so that it can be up to date in case R changes.

To disable duration validation, the builtin `.validate-duration` plugin
can simply be unloaded.