kvs: always store empty dir to content store
Problem: The kvs has always assumed an empty directory treeobj
has been stored in the content store, because that is how the first
kvs root was originally initialized.  The introduction of kvs dump / restore
can prune non-referenced objects in the content store, such as the
empty directory treeobj.  This can lead to errors.

Solution: Always store an empty directory object into the content
store when loading the kvs module.  Previously, this was only done when
a checkpoint was not found.

Fixes #4222