broker: unblock core-dump generating signals
Problem: the broker installs a signal handler for SIGSEGV and
SIGFPE and blocks other signals that generate core dumps, but we
would like a core dump immediately (the default action) when any
of those happen.

Allow default handlers to remain in effect for these signals.
This increases the odds of producing a core dump which can be
used to determine the root cause.

Add test coverage for SIGSEGV.

Fixes #4230