broker: don't log drops from unknown uuid
Problem: if a system instance broker with TBON children crashes
and restarts, it may log many messages about dropped downstream
messages from unknown uuid as it processes its peers that have
not learned of its demise.

Example:
  Mar 16 17:39:03 picl0 flux[65183]: broker.err[0]: DROP downstream
    request topic log.append from 60dcf4ce-5451-4176-abf7-9fa30270a6f6:
    unknown uuid

Suppress these messages as this is expected and the added log noise
is not especially helpful.