librlist: add void * arg to rlist_copy_internal()
Problem: The rlist_copy_internal() implementation doesn't allow a
void * argument to passed along to the underlying rnode copy function,
but this will be required in order to filter objects based on
constraints like property or hostname.

Add an arg parameter to rlist_copy_internal() and update all current
callers.