librlist: add rlist_copy_constraint()
Problem: There is no way to filter resources that match a given
constraint from an rlist object, but this may be necessary to support
testing and matching subsets of resources based on properties.

Add rlist_copy_constraint() which will return a copy of all resources
that satisfy a constraint conforming to RFC 31.