librlist/test: add property unit tests for librlist
Problem There are no unit tests that exercise overall property support
in librlist.

Add a set of simple tests for librlist property support.

Since it is challenging to construct R objects and property queries
in C, most functionality tests will be done in higher levels of
the testsuite, e.g. in tests for flux-R and Python ResourceSet
implementation.