python: validate RFC 31 constraints in Jobspec class
Problem: The Jobspec class does not do even high level validation of
RFC 31 constraints in Jobspec when constraints are specified.

Add first level validation of constraints in the Jobspec Python class.
Since RFC 31 is flexible in allowed operations and their "arguments",
just do a best effort validation which catches obvious errors. It is
intended that the scheduler actually applying constraints will do a
full validation which can be triggered in a feasiblity RPC.