sched-simple: Support RFC 31 constraints in jobspec
Problem: The simple test scheduler in flux-core does not support
the job constraints via the RFC 31 constraints object in jobspec,
but support for these constraints will be instrumental in testing
end-to-end functionality.

Add simple support for constraints by copying the main resource list so
that it only contains resources that satisfy the provided constraint,
then attempt to allocate resources from the result.