librlist: add property support to rnode structure
Problem: The librlist implementation does not currently support
properties set on resource nodes, but it is a near term requirement
to support basic, per-rank properties in flux-core.

Start by adding generic property support to the `struct rnode` class
in librlist. Implemented as a simple hash of names registered with
each rnode.