librlist: support encode/decode of RFC 20 properties
Problem: RFC 20 describes an optional properties key in the Rv1
object, but librlist, flux-core's interface to Rv1, does not yet
support this key.

Add support for the properties key when decoding an Rv1 object to
an rlist structure, and when encoding an rlist with properties to
its Rv1 JSON representation.