flux-shutdown: add new command
Problem: there is no command front end to drive instance
shutdown.

Add flux-shutdown, a command that can trigger shutdown and monitor
log activity during the shutdown process.

If a TARGET option is present, it is resolved to the URI of a Flux
instance that will be shut down.  Otherwise the current instance is
the target.

Logs are displayed until the GOODBYE state is reached.  The logging level
may be adjusted with the --quiet and --verbose options.  Alternatively,
the --background option may be use to to skip the monitoring and exit
immediately.

Fixes #3895.