broker/shutdown: add external shutdown mechanism
Problem: instance shutdown currently occurs upon completion of the
initial program or receipt of signals, but sometimes it may be
useful to have manual control over the shutdown process.

Implement a shutdown.start RPC, for rank 0 only, that interfaces with a
front end command to initiate and monitor instance shutdown.  The same
procedure is followed as when the broker catches a terminating signal:
by killing the intial program with SIGTERM, or if there is no initial
program, posting the "rc2-abort" event.  The broker transitions out of
RUN state and works through its teardown process.

If the shutdown.start RPC has the streaming flag set, log messages from
that point are forwarded to the client for display during the shutdown
process.

After rc3 completes and the broker enters the new GOODBYE state, if
there is an active shutdown client, the stream of shutdown.start
responses are terminated with ENODATA.  Upon the client's disconnect,
the "goodbye" event is posted, causing the broker to transition to EXIT
state, which unloads the connector-local module and eventually exits.

If the instance is running under systemd, the broker exit code will
be the special one that inhibits systemd from restarting the broker.