libsdprocess/strv: use calloc correctly
Problem: An array was created using calloc(), but for some
silly reason, the array length was not passed into the first
parameter of calloc().  The full length of array was calculated.

Solution: Use calloc() correctly by passing in the array length
in the first parameter.  Update variable types as needed and helper
functions as needed.