job-ingest: add constraints attribute to jsonschemas
Problem: RFC 14 defines a constraints attribute for placement of RFC
31 conformant job constraints, but this is not reflected in jobspec
jsonschema documents.

Add the constraints system attribute to both jobspec.jsonschema and
jobspec_v1.jsonschema. Currently these just validate that the given
constraints are of type object, as a more detailed schema definition
for RFC 31 constraints would be quite onerous.

Instead it is left up to the scheduler, which must apply the defined
job constraints, to validate that the constraints object is correct
and feasible.