broker: don't publish shutdown event message
Problem: the broker publishes a "shutdown" event message on
entry to the SHUTDOWN state, but it is not used.

The shutdown message is no longer used to drive the state machine,
and the state-machine.monitor streaming RPC is available for any
subsystems that need to be notified when shutdown is occurring.

Remove publishing code.
Update markdown broker state machine documentation.