flux-mini: add --requires option
Problem: While attributes.system.constraints is now supported in v1
jobspec, there is no easy way to set a valid constraints object when
submitting jobs with the flux-mini based commands.

For now, add a very simple `--requires=LIST` option to all flux-mini
commands. This simply sets the possibly comma-separated list of
property strings in a `attributes.system.constraints.properties` array.

In the future, this option can be expanded to support a more powerful
syntax which could expand directly into an RFC 31 constraints object,
but for now we start with the simplest solution.