libflux: fix sender matching for broker loopback
Problem: flux_msg_route_match_first() returns false when either
or both messges lack routing frames.  Consequently, flux_msglist_cancel()
does not work on requests sent by the broker to itself.

Modify the logic in this function so that it returns true when neither
message has routing frames, which is consistent with the intent of
the function:  to match requests that have the same sender, despite the
perhaps poorly chosen function name.

Add unit tests.