python: add require-instance job validator
Problem: It might be useful for a system instance to reject jobs which
are not themselves new instances of Flux, i.e. either batch jobs or
jobs that run `flux start` or `flux broker`. However, Flux does not
provide a way to do this.

Add a very simple `require-instance` job validator which attempts to
reject all jobs that do not create a new instance of Flux. It does
this by looking for either the "batch" system attribute in jobspec
(which will cause the job shell to start a new instance), or a command
that starts with "flux start" or "flux broker".

If necessary, this plugin could be copied and modified in environments
that want to be slightly more permissive, or allow other command
arguments that might result in a new instance. However, at least Flux
will provide an example from which to base new validators.

Fixes #4214