systemd: make directory init failure non-fatal
Problem: the content.sqlite file is only created on rank 0 but if
the systemd unit file ExecStartPre fails to set the owner and mode
of ${localstatedir}/lib/flux on any rank, then flux fails to start.

Example: ${localstatedir}/lib/flux is a symbolic link pointing to
/scratch/flux on all ranks, but /scratch is only mounted on rank 0.
Flux fails to start on the nonzero ranks.

Add a "-" in front of the commands so that the unit succeeds even
if those commands fail.