flux.service: use StateDirectory for content.sqlite
Problem: the flux service unit file currently creates, chowns,
and chmods ${localstatedir}/lib/flux for the content.sqlite file
on startup using ExecStartPre, but systemd provides support for
standardized creation of a directory for persistent state.

One notable change is that the old directory was relative to
flux's --prefix and the new one is relative to systemd's.
This will have no effect on system-packaged flux which is the
main use case for execution under systemd.

Also of note: the directory may not be a symbolic link.
If content.sqlite needs to be redirected to a different file
system, a bind mount on /var/lib/flux may be used for now;
in the future, we could let this path be configured via TOML.

Since ${STATE_DIRECTORY} was introduced in systemd v240 and
RHEL 8 is currently using v239, add a workaround to substitute
a default value if ${STATE_DIRECTORY} is not set.

Fixes #4081