broker: don't leave shutdown state prematurely
Problem: broker immediately posts "children-none" on entry
to STATE_SHUTDOWN, even if it has TBON children.

state_machine->child_count is always set to zero on entry to
STATE_SHUTDOWN because it hasn't been updated since it was initialized,
when no peers were online. There is actually no reason to be saving this
value - just call the overlay_get_child_peer_count() accessor when the
count is needed instead of caching it.

Add a failing test to the broker state machine sharness test script.