librlist: support exclusive node allocations
Problem: Jobspec v1 supports an exclusive flag for node exclusive
allocations, but rlist_alloc() does not have any support for an
exclusive flag.

Add an exclusive flag to `struct rlist_alloc_info`. Support exclusive
node allocation in rlist_alloc() via the new flag, but only when
there is an explicit request for a number of nodes.