python: add exclusive flag to Jobspec constructors
Problem: Jobspec from_*_command() constructors do not have a way to
specify the exclusive flag to request nodes exclusively.

Add an exclusive flag to all constructors, which defaults to False.
Since the exclusive flag must be specified on a node resource vertex,
raise an error if the exclusive flag is requested without also
specifying an explicit node count.