sched-simple: support exclusive flag
Problem: The jobspec reader in sched-simple supports reading an
exclusive node flag from jobspec, but sched-simple doesn't pass this
flag to the librlist allocator, so effectively the exclusive flag is
not supported.

Pass the exclusive flag from jobspec to rlist_alloc() in sched-simple
so that exclusive node requests are supported.