broker: support special norestart exit code
Problem: need a way for systemd to tell Flux what broker exit
code to use to inhibit automatic restart.

If the broker attribute 'broker.exit-norestart' is set, store it's
value to be used in situations when automatic restart is not desired.
The value is set to zero if the attribute is not set (e.g. broker
was not started by systemd so there is no implicit restart).