broker: prevent systemd restart if rc1 fails
Problem: if rc1 fails in a system instance, the broker is
automatically restarted by systemd, but the main anticipated
cause of rc1 failure is misconfiguration, so a more appropriate
response is to leave Flux off until someone can resolve the
problem and restart it manually.

If rc1 fails, and the "broker.exit-norestart" attribute is set,
exit with that value instead of the script exit code.