content-files,sqlite: use statedir not content.backing-path
Problem: content persistence semantics are confusing

Instead of having each content backing module implement persistence
semantics based on their interpretation of the content.backing-path
attribute (a directory in content-files, a file in content-sqlite),
simply make the content persistent if statedir is defined.

In the case of content-sqlite:  if statedir is defined, the backing
file is $statedir/content.sqlite, otherwise it is $rundir/content.sqlite,
and we know that statedir is persistent and rundir is not.

This means the basename of the content.sqlite file or the content-files
directory is hardwired - just the directory location changes.
If this is too inflexible, in the future we could make it configurable,
but for now it seems fine.

Update sharness tests.