librlist: add rlist_properties_encode()
Problem: There is no way to query all properties for a struct rlist,
except by encoding the rlist to an R object and looking for a
properties key.

Add rlist_properties_encode() which returns a JSON string that encodes
all properties for a struct rlist using the specification defined
in RFC 20. This gives callers more convenient access to the list of
properties defined in a resource set, without first encoding to R.