broker: check that instance owner owns directories
Problem: the broker checks that rundir, statedir, and the directory
containing the local uri are u+rwx, but it does not check that the
owner of the directory is the instance owner.

Check that st_uid == getuid ().

Update the local-uri override test in t0001-basic.t to use the
trash directory instead of /tmp to contain the test socket.