python: preserve ResourceSet state in operations
Problem: It is surprising when the ResourceSet special "state"
property is cleared on copy, union, diff, intersect, etc.

Preserve the state in the copy() and all other "operator" methods. The
preserved state is always from the "self" parameter, i.e. the source
resource set, which may not always make the most sense, but is probably
the least surprising solution here.