t0014-runlevel.t: run interactive shell under pty
Problem: The test 'default initial program in $SHELL' in
t0014-runlevel.t executes the interactive shell without a tty,
which may break in future flux-core versions that issue an error in
this situation.

Run the test under runpty.py to avoid the potential failure. Use the
runpty --input=none option instead of redirecting from /dev/null.

In case of a hang due to problems with sending EOF over the pty, also
execute the test under run_timeout(), adding SHELL=/bin/sh to the
--env option of this program.

Since the runtpy.py script must be able to load the flux Python module,
execute this script under flux(1) so that paths are set up correctly,
otherwise the Python modules may not be found.