testsuite: fix test that runs interactive shell
Problem: A test in t2212-job-manager-plugins.t runs an interactive
shell.  This works even without a tty since the test expects a failure,
but soon the broker may exit with a different error in this situation,
breaking the test.

Add /bin/true to the arguments in the test so that the shell is no
longer interactive. This doesn't change the semantics of the test,
but will prevent future failures if the broker checks for a tty.