broker: fail if no tty and rc2 is interactive shell
Problem: It is a common folly to run `flux start` via flux-mini run
or foreign RM launcher and forget to allocate a pty to the job. This
can waste time as a user waits for a shell prompt that will never
come.

Since it is almost guaranteed that running an interactive shell without
a pty is a mistake, catch this condition in the broker and emit a useful
error message and exit.

Now, `flux mini run flux start` without `-o pty` or `-o pty.interactive`
will result in the error:

 flux-broker: stdin is not a tty - can't run interactive shell
 flux-job: task(s) exited with exit code 1

Fixes #2032