testsuite: reopen runpty stdio with surrogateescape
Problem: Python has issues with utf-8 sequences when LC_ALL=C is set
in the environment (which it is for the testsuite). Even when we're
very careful about encoding data with utf-8 and errors=replace, when
the data is written to stdout/err, Python will senselessly raise an
exception and abort.

Re-open stdout and stderr with encoding="utf-8" and
errors="surrogateescape" to work around this problem, even when the
current LANG is set to C.