broker: add overlay_msg_is_local()
Problem: Broker services may need to determine if a request originated
locally in order to block remote requests for security purposes,
but it is not straightforward to make this determination since routes
are encoded as uuids not ranks.

Tag messages that are received from a parent or child with a special
key in the flux_msg_t aux hash, and use the presence of this key in
a new overlay_msg_is_local() function to determine that a message
was receieved remotely, and therefore is not local.