libsubprocess: add auth callback to subprocess server
Problem: There is no way for a user of a subprocess server to reject
exec requests before they are turned into local processes, but this
may be necessary to implement security requirements.

Add an optional "authorization callback" to the subprocess server.
This function will be called for each exec request, and allows the
implementation to return -1 (with errno set) if the request should
be rejected with an error.