broker: reject remote exec requests on rank 0
Problem: There is no way to prevent root or flux from running arbitrary
commands (as flux) on the node running the rank 0 broker. If that
happens to be a cluster management node, some sites have a policy of
restricting access to limit the effect of a compromised compute or
login node.

Establish a subprocess server authorization callback on rank 0 that
denies all exec requests by default that do not originate locally.

This is presumed safe because the main use cases for remote subprocess
execution all originate on rank 0 anyway (flux exec, the job-exec
service, etc.), and this step directly addresses the specific problem
statement above.

Configuration of this policy can come later.

Fixes #3202