broker: abort rc scripts with SIGHUP not SIGTERM
Problem: the broker sends SIGTERM to the commands it spawns for
rc1, rc2, cleanup, and rc3, but most of these are shell scripts and
SIGTERM is ignored by bash when a subcommand is running.

SIGHUP has the same default action as SIGTERM but is forwarded by
bash to subprocesses.  Change runat_abort() to use SIGHUP instead
of SIGTERM.