job-archive: require period, set default dbpath
Problem: With the new statedir broker configuration, a number of
modules like content-sqlite automatically store content in the
statedir (or rundir as backup).  However, the job-archive module
still requires users to configure a database path.

In addition, the job-archive would not run if a dbpath was not
set by the user.  If we set a dbpath by default, that means the
job-archive will always run when loaded.

Solution: If a dbpath is not configured, assume the database should
be stored in $statedir/job-archive.sqlite.  If statedir is not set,
return error.

As as side effect, require the user to always set the job-archive
period.  This will indicate to the job-archive module that it should
load the database and populate it.

Update / add tests accordingly.