etc/rc1: only load job-archive when appropriate
Problem: The job-archive module has always been loaded in rc1.
Previously, if a database path was not specified, the module simply
did nothing.  However, recent changes to job-archive would default
to database path in 'statedir' if it was set.  This could lead to
job-archive loading / running too many job-archives.  In addition,
an error will occur if period is not set when loading job-archive.

Solution: Only load the job-archive module if we are on rank 0 and
the job-archive period configuration has been set.

Fixes #4251