libjob: allow a count of 0 for gpu resources
Problem: The Python jobspec v1 class allows a count of 0 for
resources that aren't node, slot or core, but the jobspec1 C API
rejects a jobspec with any resource count of 0 as invalid.

Allow gpu resources specifically to have a count of 0 since this
is the only resource supported by the jobspec1 class besides node,
slot and core.