ingest: update schemas to allow resource count of 0
Problem: The Python and libjob jobspec1 implementations both
have been relaxed to allow a count of 0 for some resource types,
but the schemas in src/modules/job-ingest/schemas fail to validate
newly allowed jobspec.

Update the jobspec jsonschema to allow a count 1 of resources other
than "node", "slot", and "core" to match the other jobspec validators
in flux-core.