job-manager: add reconfig callback capabilitiy
Problem: configuration updates are only processed by the job manager
as updates to the config object cached in the flux_t handle.  There
is not a way for job manager subsystems to register handlers for config
changes.

Add a mechanism for subsystems to register callbacks that are invoked
immediately with the current config object, and depending on their
initial return value, later each time the configuration changes.