job-manager: purge inactive jobs
Problem: the overhead of inactive jobs grows without bound in a
system instance.

Add a mechanism to trim the oldest inactive jobs from the job manager
internal hash and KVS.  There are two ways to do it.  One way is to
make a job-manager.purge RPC specifying either a maximum number of
retained inactive jobs, or a maximim age (time since transition to
inactive) of retained inactive jobs, or both.

The other way is to configure those criteria via TOML, e.g.

  [job-manager]
  inactive-age-limit = "7d"
  inactive-num-limit = 10000

The KVS transactions used to unlink batches of jobs is limited to 100
entries to avoid giant messages and head of line blocking in the broker.
When purging via configuration, this limits the number of jobs purged per
heartbeat interval.  When purging via RPC, the caller specifies the batch
size (up to the maximum) and must check the returned job count, and repeat
the RPC until a count less than the batch size is returned.

Note: KVS job.* directories that become emtpy after all the jobs in them
have been purged are NOT removed by the purge.  This can be accomplished
later by an offline dump/restore of the content backing store, since empty
directories are not included in dumps.