libterminus: stash streaming rpc future in client struct
Problem: The flux pty client code issues a streaming RPC request to
attach to a remote pty, but the RPC future is not stored in the pty
client structure, and so is not available to destroy or cancel if
the client needs to be shut down early, e.g. before the first response.

Stash the RPC future in the struct flux_pty_client.