flux-job: stash struct flux_pty_client in attach_ctx
Problem: When flux-job attach needs to attach to an interactive tty,
the resulting pty client structure is not added to the attach context,
and thus can not be referenced from functions other than the pty
client callbacks. This can prevent necessary cleanup when a job
exception occurs, for example before the client pty is fully connected.

Stash the `struct flux_pty_client` in `struct attach_ctx` so it is
available from all or most job attach callbacks.