libterminus: attach pty client immediately in synchronous mode
Problem: When FLUX_PTY_CLIENT_ATTACH_SYNC is used on attach for a
pty client, the call to pty_client_attached() to "start" the pty is
delayed until the first reactor callback. Not only does this seem
unnecessary, it seems to violate the purpose of a "synchronous"
attach, where a caller would expect the pty to be fully attached
after the call to flux_pty_client_attach() with this flag set.

Call pty_client_attached() and reset the RPC future immediately when
the attach response is received via synchronous flux_rpc_get()
in flux_pty_client_attach().