flux-job: use asynchronous pty attach in flux-job attach
Problem: When flux-job attach needs to attach to a remote pty it uses
the FLUX_PTY_CLIENT_ATTACH_SYNC flag, which causes the program to
block waiting for a response from the remote pty. Meanwhile, flux-job
attach is blocked from processing other job events, most notably job
exceptions. It would be better to asynchronously attach to the remote
pty so these events can be processed and appropriate actions taken.

Remove the FLUX_PTY_CLIENT_ATTACH_SYNC flag from flux-job attach.

Since the attach is now asynchronous, remote the return code from
attach_pty() since its only use case was to determine when the remote
pty no longer exists.

In pty_client_exit_cb(), check to see if the pty was attached when
it "exited". If it was not, then assume the exit was really an
error during attach, and fall back to setting up "normal" KVS based
stdio. This allows `flux job attach` to work for completed jobs which
had an interactive pty.