flux-job: terminate unattached pty on job exception in attach
Problem: If a job exception occurs early in flux-job attach, and there
is an interactive pty, a hang may result while the program waits for a
response from the remote pty server which will never come.

Terminate a pty client immediately if it is not yet attached to the
server and there was a fatal job exception.

Fixes #4281