testsuite: fix intermittent failures in cron faketime test
Problem: The 'flux-cron tab works for set minute' fails intermittently,
especially after updating the copy of libev in flux-core to 4.33.

Since all other tests in t0016-cron-faketime.t use a datetime relative
to "today" instead of a specific datetime as in the failing test,
it was determined by trial and error that modifying the faketime
relative to "today" seemed to work around these sporadic failures.

It is currently unknown why this is the case, but since this is
for testing only, just accept that the adjustment seems to work and
move on.