testsuite: test empty return from flux.importer.import_plugins
Problem: There is no test that ensures the flux.importer.import_plugins
function, used by the job validator, returns a dict and not a list when
there are no plugins found.

Because there is no import_plugins() unit test, simply add a specific
test to t2110-job-ingest-validator.t which verifies this case.