content-sqlite: use safer pragmas in statedir
Problem: the system instance content.sqlite file is sometimes corrupt
after a restart.

We are unconditionally setting journal_mode=OFF, synchronous=OFF which
is trading performance for the possibility of corrupting the database.

When statedir is set, and thus the database is being preserved on exit,
change these pragmas to journal_mode=WAL, synchronous=NORMAL which is
still pretty fast, but less likely to corrupt the database on broker crash.

For testing, allow the journal_mode and synchronous values to be set
on the module command line.

Hopefully fixes #4152