flux-dmesg: Add option for human-readable output
Problem: The default output from flux-dmesg(1) is useful, but slightly
difficult for humans to read. This is especially true when trying
to parse timestamps "by eye", which are in ISO 8601 format at UTC
timezone.

Add a new -H, --human option to flux-dmesg(1) which largely mimics
the option of the same name in the Linux dmesg(1) command. Timestamps
are simplified by printing a human readable datetime string at least
once per minute, with following timestamps as offsets since the last
"timebreak". An option to print deltas between each message is also
offered via -d, --delta.

The human readable output is also colorized by default when stdout
is connected to a tty, or always if --color is used. Timestamp,
facility name, and the actual message are all given different colors
to aid readability.