shell: cleanup refs on reconnect
Problem: When a reconnect happens in the shell, eventlogger
references may never get cleared.  This can lead to the shell hanging
and not exiting when tasks are complete.  The shell will eventually
"doom" timeout as a result.

Solution: Setup "reconnect" callbacks to clear eventlogger references
whenever a reconnect occurs.  It is the only way to ensure those references
are cleared when we are unsure if a eventlogger response was received.