rc: manage dump/restore of backing store content
Problem: we need a way to tell rc scripts to restore content
on startup, and dump content on shutdown, for offline KVS garbage
collection of a system instance or user checkpoint/restart.

Add some logic to rc1 and rc3:

rc1:  If the content.restore broker attribute is set to a file path,
then load the content backing store module with the 'truncate' option,
and restore content from the file before loading the KVS.

rc3:  If the content.dump broker attribute is set to a file path,
then dump content to the file after unloading the KVS.

Additionally, if content.restore=auto, then rc1 looks for a symlink
named RESTORE in the broker's current working directory or ${statedir}
if defined.  If the symlink exists, then restore content from the file
it points to and remove the symlink on success.

If content.dump=auto, then rc3 dumps content to an automatically generated
file name containing the date in the current working directory or
${statedir} if defined, and creates the RESTORE symlink pointing to it.