content-s3: add stubbed truncate module option
Problem: rc scripts use the content backing store 'truncate' option
to manage offline garbage collection, but content-s3 does not
support this option.

Add a truncate module option.  Since emptying an s3 bucket is not
directly suported by libs3, this is a rather involved process.
For now, if this option is supplied, log an error instructing the
user to purge the bucket using s3 console or another mechanism and
return failure.

Add test.