testsuite: remove test_set_prereq in test files
Problem: In a number of test files "test_set_prereq LONGTEST" is
called when TEST_LONG is set to 't'.  This is unnecessary because
this is already done in flux-sharness.sh.  It appears this has
been copy and pasted around over time.

Solution: Remove the call to "test_set_prereq LONGTEST".