content-sqlite: use BLOB type for hash keys
Problem: CHAR(20) type for hash keys would appear to truncate
sha256 hash keys, and implies that character set encoding/collation
code is used in key comparisons.

It turns out that content-sqlite ignores the CHAR length parameter,
and that character set stuff may be effectively a no-op here - at least
simple tests in #2795 show that it may be equivalent to BLOB in terms
of stored size and comparison overhead.

Anyway, CHAR(20) looks wrong, so change it to BLOB.

Fixes #2795