python: uri: fix intermittent failure of fallback pid resolver
Problem: The fallback PID URI resolver was failing intermittently
during testing due to what was presumed whitespace in the command
string found in /proc/[pid]/stat. Since the /proc/[pid]/stat contents
were split into fields on whitespace, the incorrect field was passed
to int() and an exception thrown, resulting in the inscrutable error:

 flux-uri: ERROR: invalid literal for int() with base 10: 'S'

To fix, use a regex instead of simply splitting the contents on
whitespace. This allows whitespace to possibly occur in the command
name, since we now consume everything between parentheses as the
command name.

Fixes #4168