flux-overlay: replace status --no-color option with --color=WHEN
Problem: The flux-overlay status command requires manual disabling of
colorized output using a --no-color option, but other Flux commands
use the semi-standard --color=WHEN option.

Replace --no-color with `-L, --color=WHEN`, where WHEN can be one of
'auto', 'never', or 'always'. The default is 'auto' if the option
is not specified, or 'always' if the option is specified without
an argument.

Since --color=auto is now the default, color will not be displayed
if stdout is not connected to a tty. Therefore the use of --no-color
in the testsuite can simply be removed.