flux-overlay: Add -H, --highlight option to status subcommand
Problem: It would occasionally be useful to highlight the overlay
subtrees which connect a given set of hosts or ranks, but the
flux-overlay status subcommand does not currently include this feature.

Add a `-H, --highlight=TARGETS` option to `flux overlay status`
which will hightlight the name of ranks where the idset of ranks in
the subtree intersects with the provided target ranks or hostnames.