job-list: consistently return job attributes
Problem: The job-list module was inconsistent in its return of
job data to the user.  In some cases, "defaults" were returned
when data was never initialized.  In other cases, the data is
simply not returned to the user.

Solution: Consistently do not return data to the user when it
is not available / not set.  The attributes adjusted are:

name
ntasks
nnodes
ranks
nodelist
expiration

Adjust tests accordingly.

In the Python JobInfo class, set display defaults for name and ntasks.