shell: fall back to /tmp if TMPDIR cannot be created
Problem: If a user inadvertently exports a TMPDIR that can't
be created on the node on which the job eventually runs, then the
tmpdir job shell plugin raises a fatal error. This can occur when
the parent directory of TMPDIR doesn't exist on the system on which
the job eventually runs.

Instead of raising a fatal error, issue a warning in this case and
fall back to /tmp for TMPDIR.

Fixes #4328