t2602-job-shell.t: adjust job shell TMPDIR tests
Problem: The job shell TMPDIR tests in t2602-job-shell.t do not align
with current expected behavior of the shell tmpdir plugin.

Revise the test for a bad TMPDIR to ensure the job succeeds and TMPDIR
is reset to /tmp.

Add a test to ensure that an unset TMPDIR is preserved within a job.