shell: do not set TMPDIR to FLUX_JOB_TMPDIR when unset
Problem: The job shell tmpdir plugin sets TMPDIR to FLUX_JOB_TMPDIR
when TMPDIR is not already set. However, this causes problems if
the current job submits more jobs, because TMPDIR now points to a
job specific directory. This can cause jobs to fail when they are
executed on different nodes (where the job tmpdir does not exist).
If the job runs locally, then it shares TMPDIR with another job which
is probably not the original intent of the feature.

For these reasons, avoid setting TMPDIR to FLUX_JOB_TMPDIR. Just
leave TMPDIR unset if the user submitted the job that way.

Fixes #4329